Added a simple, crude, Mocha-based Squid bug #7 test case.
Squid bug #7 is a violation of the "Proxy MUST update previously
cached headers on 304 responses" requirements.

This test case illustrates some limitations of Mocha and our own APIs.

I tried to encapsulate common proxy-testing operations into
tests/ProxyCase, but it is not yet clear whether that was the right
approach: Most of the code in that class deals with starting and
waiting for HTTP client and server agents. There may be a better
way to encapsulate and reuse that code.

Similarly, ProxyCase::expectStatusCode() may need to be moved into
HTTP Client (or even HTTP Agent!).