Initial support for automatically starting/stopping the DUT
...(a.k.a. Device Under Test) using Proxy Overlord Protocol.

The earlier (uncommitted) implementation was adding the Overlord (i.e.
DUT controller) via the existing HttpCase::proxy() API, but that
essentially implied restarting the DUT for every test case. Sometimes
that may be a good idea, but, when dealing with tests with lots of test
cases, it is just too expensive and, hopefully, unnecessary. Also,
the HttpCase API relies on "transactions" and those are a bit awkward
to provide for a mostly black-box and often remote proxy (as opposed
to the built-in Proxy agent).

After realizing all that, I moved Overlord maintenance into the newly
added Test::startup/shutdown() methods.

The current Overlord implementation uses hard-coded Squid-specific
configuration and that configuration is somewhat specific to the
test-happy-eyeballs.js test. We will expose configuration support later
so that each test can configure the DUT to match the test needs.