Switched to FlexibleConfigGen to filter out bad generated configs
This filtering reduces the total number of configurations from 1890 to
1530 despite enabling SMP range tests and adding "fat" ranges (see
below)! We can probably filter out more. Requires SQUID-479-large-header
commit 3a2e241 (a.k.a. master commit f61a693).

Also added a "fat" Range configuration, to test a single range
containing multiple bytes (that is not the _whole_ body).

Also re-added range checks for SMP mode. We may need to disable that
combination, but we should document why we have to (if we do).