fixup: Removed unusable (and broken?) --max-prefix-size default
When Configurators() "vary" a command line option, it is probably not
possible to use the default option value: If the user does not specify
the option, then they get the values generated by Configurators().
Perhaps this is a Daft CLI flaw, but we do not need to solve that
problem right now.

The command line option expects KB but the default was in bytes AFAICT.