accumulate-headers-after-304: Adjusted reply_header_max_size defaults
When deciding on the set of reply_header_max_size values to test, I kept
in mind that the upcoming stored-header test will cover basic storage of
4KB, 16KB, and 32KB headers. In this accumulate-headers-after-304 test,
we can probably pay less attention to various Store-related magic sizes
and focus on reply_header_max_size semantics itself. This logic resulted
in the reduction of accumulate-headers-after-304 tests from 45 to 27.