tests/proxy-collapsed-forwarding: Polish body-less soLiveFeeding
While working on tests/revalidate-collapsed (which cloned this test), I
discovered that zero-size responses cannot be used for soLiveFeeding
cases because we cannot tell the server to block (after sending headers
and) before sending body when there is no body to send.

Existing code did not fail despite this flaw because it does not (and
cannot?) see that an soLiveFeeding case ended up with an soPureHits
outcome -- the entire response had been (usually?) cached by the time
hit requests arrived.

TODO: Convert ConfigGen-based code to FlexibleConfigGen so that it is
easy to skip zero bodySize() and soLiveFeeding configurations. That
upgrade is not trivial because of --retries: There is no replacement API
for ConfigGen::addGlobalConfigAdjustment() AFAICT.