tests/proxy-collapsed-forwarding: Reliability/polishing improvements
### finishStagingRequests() in soLiveFeeding

I am not sure whether waiting for staged requests would fix in any
previously unstable cases, but it makes sense to wait for that event (in
expecting-success cases) and to keep soLiveFeeding code similar to
soTrueCollapsing code that already does that.

### Do not rely on memory to save disk hits

When disk caching is enabled, Squid may purge cached responses from
memory cache or refuse to memory-cache responses to begin with. And
since a disk cache entry will still be locked for writing at hit request
time in soLifeFeeding cases, those cases cannot expect cache hits just
because cache_mem is also enabled.

### Polish body-less soLiveFeeding

While working on tests/hit-revalidation (which cloned this test), I
discovered that zero-size responses cannot be used for soLiveFeeding
cases because we cannot tell the server to block (after sending headers
and) before sending body when there is no body to send.

Existing code did not fail despite this flaw because it does not (and
cannot?) see that an soLiveFeeding case ended up with an soPureHits
outcome -- the entire response had been (usually?) cached by the time
hit requests arrived.

TODO: Convert ConfigGen-based code to FlexibleConfigGen so that it is
easy to skip zero bodySize() and soLiveFeeding configurations. That
upgrade is not trivial because of --retries: There is no replacement API
for ConfigGen::addGlobalConfigAdjustment() AFAICT. On the other hand, it
is not clear whether --retries are still needed, especially after this
commit improvements.

----

Merged proxy-collapsed-forwarding-specific changes in
SQUID-948-updatedReply-assert (as of commit 12ae6fdd).