tests/proxy-update-headers-after-304: Wait for updates to finish
This is a temporary hack until finishCaching() covers header updates.
Without this hack, this test often fails when executed under Valgrind
(with or without ALL,9 debugging), probably because it takes longer for
updates to be completed when running under Valgrind.

tests/accumulate-headers-after-304 does not need this hack because it
uses a "This is step 4, not 3" hack instead. We could use a similar
check-same-worker hack for proxy-update-headers-after-304, I guess, but
the sleep hack feels a bit better for this particular test because it
allows to test cross-worker updates.

----

Cherry-picked SQUID-1011-test-smooth-reconfiguration commit 77ec5bf2.